Safari-611.1.10/LayoutTests/editing/text-iterator/count-matches-in-form-expected.txt

PASS internals.countMatchesForText('rule', ['CaseInsensitive', 'AtWordStarts', 'TreatMedialCapitalAsWordStart', 'WrapAround'], '') is 1
PASS successfullyParsed is true

TEST COMPLETE
